stats: Reformat match block
Ostensibly pissed off at `nvim-nu` (particularly the tree-sitter-nu it
used which is unquestionably out-of-date) causing highlights to break
of in the middle of words, ie, [not]hing, until `nushell/tree-sitter-nu`
came along which supported more recent versions!

Still a little uneasy with the highlights having been used to built-in
highlights for languages like Bash and Fish. Hopefully this will be
fixed soon (see nushell/tree-sitter-nu#99 and nushell/tree-sitter-nu#98).